+++ app/src/main/java/alexander/voronov/gitapp/domain/UserEntity.kt
@@ -1,5 +1,6 @@
-package alexander.voronov.gitapp
+package alexander.voronov.gitapp.domain
 
+//базавая сущность всего приложения
 data class UserEntity(
     val login: String,
     val id: Long,
