+++ app/src/main/java/alexander/voronov/gitapp/domain/UsersRepo.kt
@@ -1,8 +1,10 @@
-package alexander.voronov.gitapp
+package alexander.voronov.gitapp.domain
+
+import alexander.voronov.gitapp.domain.UserEntity
 
 interface UsersRepo {
     //CRUD в данном случае только Read
-    //асинхронный подход. метод ожидаеь на вход 2 callback 'a
+    //асинхронный подход. метод ожидает на вход 2 callback 'a
     fun getUsers(
         onSuccess: (List<UserEntity>) -> Unit,
         onError: ((Throwable) -> Unit)? = null
