+++ app/src/main/java/alexander/voronov/gitapp/ui/users/UserViewHolder.kt
@@ -1,5 +1,7 @@
-package alexander.voronov.gitapp
+package alexander.voronov.gitapp.ui.users
 
+import alexander.voronov.gitapp.R
+import alexander.voronov.gitapp.domain.UserEntity
 import alexander.voronov.gitapp.databinding.ItemUserBinding
 import android.view.LayoutInflater
 import android.view.ViewGroup
