+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:gravity="center"
     android:orientation="vertical"
     android:padding="@dimen/activity_main_padding"
-    tools:context=".MainActivity">
+    tools:context=".ui.users.MainActivity">
 
     <com.google.android.material.button.MaterialButton
         android:id="@+id/activity_main_refresh_button"
